feat: combine serde package
