+++ clients/node/client-rds-data-node/package.json
@@ -37,9 +37,8 @@
     "@aws-sdk/hash-node": "^0.1.0-preview.5",
     "@aws-sdk/invalid-dependency": "^0.1.0-preview.1",
     "@aws-sdk/middleware-content-length": "^0.1.0-preview.5",
-    "@aws-sdk/middleware-deserializer": "^0.1.0-preview.1",
+    "@aws-sdk/middleware-serde": "^0.1.0-preview.1",
     "@aws-sdk/middleware-user-agent": "^0.1.0-preview.1",
-    "@aws-sdk/middleware-serializer": "^0.1.0-preview.5",
     "@aws-sdk/middleware-stack": "^0.1.0-preview.6",
     "@aws-sdk/node-http-handler": "^0.1.0-preview.6",
     "@aws-sdk/protocol-rest-json": "^0.1.0-preview.5",
