+++ packages/middleware-serde/src/index.ts
@@ -1,10 +1,16 @@
 import {
+  RequestSerializer,
+  SerializeHandler,
+  SerializeHandlerArguments,
+  SerializeMiddleware,
+  SerializeHandlerOutput,
   ResponseDeserializer,
   DeserializeHandler,
   DeserializeHandlerArguments,
   DeserializeMiddleware,
   DeserializeHandlerOutput,
   Protocol,
+  SerializerUtils,
   DeserializerUtils,
   InjectableMiddleware
 } from "@aws-sdk/types";
@@ -46,5 +52,40 @@ export function deserializerPlugin<OutputType>(
     middleware: deserializerMiddleware(config, serializer),
     step: "deserialize",
     tags: { DESERIALIZER: true }
+  };
+}
+
+export function serializerMiddleware<
+  Input extends object,
+  Output extends object
+>(
+  options: SerializerMiddlewareConfig,
+  serializer: RequestSerializer<any>
+): SerializeMiddleware<Input, Output> {
+  return (
+    next: SerializeHandler<Input, Output>
+  ): SerializeHandler<Input, Output> => async (
+    args: SerializeHandlerArguments<Input>
+  ): Promise<SerializeHandlerOutput<Output>> => {
+    const request = options.protocol.serialize(serializer, args.input, options);
+    return next({
+      ...args,
+      request
+    });
+  };
+}
+
+export interface SerializerMiddlewareConfig extends SerializerUtils {
+  protocol: Protocol<any, any>;
+}
+
+export function serializerPlugin(
+  config: SerializerMiddlewareConfig,
+  serializer: RequestSerializer<any>
+): InjectableMiddleware {
+  return {
+    middleware: serializerMiddleware(config, serializer),
+    step: "serialize",
+    tags: { SERIALIZER: true }
   };
 }
