--- packages/middleware-serializer/.gitignore
@@ -1,8 +1,0 @@
-/node_modules/
-/build/
-/coverage/
-/docs/
-*.tsbuildinfo
-*.tgz
-*.log
-package-lock.json
