--- packages/middleware-serializer/.npmignore
@@ -1,17 +1,0 @@
-/src/
-/coverage/
-/docs/
-tsconfig.test.json
-*.tsbuildinfo
-
-*.spec.js
-*.spec.d.ts
-*.spec.js.map
-
-*.mock.js
-*.mock.d.ts
-*.mock.js.map
-
-*.fixture.js
-*.fixture.d.ts
-*.fixture.js.map
