--- packages/middleware-serializer/CHANGELOG.md
@@ -1,53 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [0.1.0-preview.5](https://github.com/aws/aws-sdk-js-v3/compare/@aws-sdk/middleware-serializer@0.1.0-preview.2...@aws-sdk/middleware-serializer@0.1.0-preview.5) (2019-09-09)
-
-
-### Features
-
-* commit all clients ([#324](https://github.com/aws/aws-sdk-js-v3/issues/324)) ([cb268ed](https://github.com/aws/aws-sdk-js-v3/commit/cb268ed))
-
-
-
-# 0.2.0 (2019-07-12)
-
-
-### Features
-
-* add npm badges for individual packages ([#251](https://github.com/aws/aws-sdk-js-v3/issues/251)) ([8adc10c](https://github.com/aws/aws-sdk-js-v3/commit/8adc10c))
-* update jest v20 to v24 ([#243](https://github.com/aws/aws-sdk-js-v3/issues/243)) ([1e156ab](https://github.com/aws/aws-sdk-js-v3/commit/1e156ab))
-
-
-
-# 0.1.0 (2019-04-19)
-
-
-
-
-
-# [0.1.0-preview.4](https://github.com/aws/aws-sdk-js-v3/compare/@aws-sdk/middleware-serializer@0.1.0-preview.2...@aws-sdk/middleware-serializer@0.1.0-preview.4) (2019-07-12)
-
-
-### Features
-
-* add npm badges for individual packages ([#251](https://github.com/aws/aws-sdk-js-v3/issues/251)) ([8adc10c](https://github.com/aws/aws-sdk-js-v3/commit/8adc10c))
-* update jest v20 to v24 ([#243](https://github.com/aws/aws-sdk-js-v3/issues/243)) ([1e156ab](https://github.com/aws/aws-sdk-js-v3/commit/1e156ab))
-
-
-
-# 0.1.0 (2019-04-19)
-
-
-
-
-
-# [0.1.0-preview.3](https://github.com/aws/aws-sdk-js-v3/compare/@aws-sdk/middleware-serializer@0.1.0-preview.2...@aws-sdk/middleware-serializer@0.1.0-preview.3) (2019-04-19)
-
-**Note:** Version bump only for package @aws-sdk/middleware-serializer
-
-# [0.1.0-preview.2](https://github.com/aws/aws-sdk-js-v3/compare/@aws-sdk/middleware-serializer@0.1.0-preview.1...@aws-sdk/middleware-serializer@0.1.0-preview.2) (2019-03-27)
-
-**Note:** Version bump only for package @aws-sdk/middleware-serializer
