--- packages/middleware-serializer/README.md
@@ -1,6 +1,0 @@
-# @aws-sdk/middleware-serializer
-
-[![NPM version](https://img.shields.io/npm/v/@aws-sdk/middleware-serializer/preview.svg)](https://www.npmjs.com/package/@aws-sdk/middleware-serializer)
-[![NPM downloads](https://img.shields.io/npm/dm/@aws-sdk/middleware-serializer.svg)](https://www.npmjs.com/package/@aws-sdk/middleware-serializer)
-
-Provides the middleware serializer for the AWS SDK
