--- packages/middleware-serializer/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@aws-sdk/middleware-serializer",
-  "description": "Provides the middleware serializer for the AWS SDK",
-  "version": "0.1.0-preview.5",
-  "scripts": {
-    "prepublishOnly": "tsc",
-    "pretest": "tsc -p tsconfig.test.json",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@aws-sdk/types": "^0.1.0-preview.5",
-    "tslib": "^1.8.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^24.0.12",
-    "jest": "^24.7.1",
-    "typescript": "~3.4.0"
-  },
-  "main": "./build/index.js",
-  "types": "./build/index.d.ts",
-  "author": {
-    "name": "AWS SDK for JavaScript Team",
-    "email": "",
-    "url": "https://aws.amazon.com/javascript/"
-  },
-  "license": "Apache-2.0"
-}
