--- packages/middleware-serializer/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "declaration": true,
-    "strict": true,
-    "sourceMap": true,
-    "downlevelIteration": true,
-    "lib": [
-      "es5",
-      "es2015.promise",
-      "es2015.collection",
-      "es2015.iterable",
-      "es2015.symbol.wellknown"
-    ],
-    "rootDir": "./src",
-    "outDir": "./build",
-    "importHelpers": true,
-    "noEmitHelpers": true,
-    "incremental": true
-  }
-}
