--- packages/middleware-serializer/tsconfig.test.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "sourceMap": false,
-    "inlineSourceMap": true,
-    "inlineSources": true,
-    "rootDir": "./src",
-    "outDir": "./build",
-    "incremental": true
-  }
-}
